left: M7, M8, M9

+++ M1/ex01/srcs/ZombieHorde.cpp
@@ -6,7 +6,7 @@
 /*   By: mortega- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/18 23:22:29 by mortega-          #+#    #+#             */
-/*   Updated: 2022/09/18 11:54:14 by test             ###   ########.fr       */
+/*   Updated: 2023/01/29 11:30:19 by mortega-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ Zombie	*zombieHorde(int N, std::string name)
 {
 	Zombie	*horde = new Zombie[N];
 
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		horde[i].setName(name + std::to_string(i));
 	return (horde);
 }
